# feat: Improved latency for executeStreaminSql calls