--- a/Spanner/composer.json
+++ b/Spanner/composer.json
@@ -7,7 +7,7 @@
         "php": "^8.0",
         "ext-grpc": "*",
         "google/cloud-core": "^1.52.7",
-        "google/gax": "dev-main as 1.32"
+        "google/gax": "^1.32"
     },
     "require-dev": {
         "phpunit/phpunit": "^9.0",