--- a/composer.json
+++ b/composer.json
@@ -50,7 +50,7 @@
         "monolog/monolog": "^2.9||^3.0",
         "psr/http-message": "^1.0|^2.0",
         "ramsey/uuid": "^4.0",
-        "google/gax": "^1.30",
+        "google/gax": "dev-main as 1.32",
         "google/common-protos": "^4.4",
         "google/auth": "^1.34"
     },